Group endpoints by origin
This is a pretty big change.  It changes the reporting cache so that
there's a single "client" entry for each origin.  That client contains a
list of "endpoint groups", each of which contains a list of "endpoints".
These are the same objects as before, just organized differently.

More importantly, this also changes the format of the JSON array in the
`Report-To` header:

  - Each element in the array now describes an endpoint group, not an
    individual endpoint.

  - The `max-age` and `include-subdomains` properties now belong to
    endpoint groups, and not to individual endpoints.