Clarify name of "observability" property
In w3c/network-error-logging#77 there was some confusion about what
"observable from JavaScript" means.  The intended meaning is very
specific: the report will not be handed off to any registered
`ReportingObserver`s.  Another reading that is you can perform whatever
complicated analysis you want, and somehow detect that a report was
created.  This patch renames the term to something that hopefully can
only be read in the more specific way.