Clarify CORS behavior for report uploads (#97)
Report uploads are supposed to behave The Right Way and send CORS
preflight requests.  We have to set Fetch's `unsafe-request` flag to
trigger the preflights, and we also need to have a reasonable `Origin`.
We use the origin of the reports themselves.  This means that any
particular report upload can only contain reports about a single origin,
so this patch also adds an extra step to the upload algorithm to sort
reports by origin after endpoints have been chosen.