Allow report types to customize the report URL
Before, a report's URL was always the URL of the containing page (or
more accurately, of the environment settings object used to queue the
report).  For some report types, the containing page isn't relevant.
For instance, a NEL report describes a network request independent of
where or why the request was made.  A NEL report's URL should really be
the URL of the resource that was requested.

This patch allows report types to provide a custom URL if appropriate.
The document URL is a good default, though, so this customization is
optional.