Clean up subdomain match logic
Subdomain matching applies to domain names, but we were applying the
logic to origins.  Not all origins have domain names!  This cleans up
the text to make it more precise.  A nice side effect is that it's now
more well-defined what should happen e.g. to `include_subdomains`
policies for requests to a raw IP address.