Remove persistent non-document reporting from spec.
This change removes the concepts which are only relevant to persistent
out-of-band reporting specs, such as Network Error Logging, and
explicitly defines reporting mechansims for documents, where the
generated reports are queued with the document itself, and thereby
inherit its lifetime. Generic reporting concepts are extracted into
another section so that they can be used by other specs.

Addresses issues raised in #158.