Add support for workers (#206)
* Add support for workers

This adds support for configuring reporting in worker scripts, and for
generating reports from a worker context. Initialization of reporting endpoints
is decoupled from fetch so that it can be referenced in HTML instead, when
documents and workers are created. Endpoint lists, reports, and observers
are moved from Document to the relevant global scope.

Fixes: #208, #205