Change Report and ReportBody interfaces to dicts.
This change removes any suggestion that the two interface names (and
the names of those interfaces which inherit from `ReportBody`) should
be exposed on the global object. `ReportBody` is purely abstract, and
exists so that it can be used as the type of the `body` member of
`Report`, while still allowing other specifications to define new
subtypes for their own report types.

Closes: #216